[iOS] Add description for ItemIdentifiers
In some crashes the item identifiers are declared equals but the only
trace is their URL. As this is not what is used to check the equality,
add a custom description to have better debug traces.

Low-Coverage-Reason: HARD_TO_TEST this is used for debug/crash reporting.
Bug: 1506636
Change-Id: Iac69f37630c1241824ecd379d02342660e48374b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5149650
Reviewed-by: Louis Romero <lpromero@google.com>
Commit-Queue: Gauthier Ambard <gambard@chromium.org>
Auto-Submit: Gauthier Ambard <gambard@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1240579}