[iOS] Use DiffableDataSource for PageInfo table view
This CL updates the PageInfo, to use the DiffableDataSource API
instead of the custom Chrome API.
This change include:
- Using translation between the enum and NSNumber
- Register/dequeue the cells manually
- Handle the header/footer manually
- Handle the cells manually

The main functional change is that the data are now "pulled" when
displayed instead of being stored in the item (i.e. when the Table
View wants to display the cell we look at the value from the
permissionsInfo instead of looking at the permissionsInfo at the
creation time).

This CL is also adding helpers to register/dequeue cell and headers
more easily.
It is also creating a new base class to be able to handle the
empty view and other things if needed in the future.

But on the other end, those changes mean that:
- Reloading the cell is much easier: as the information are looked up
  when the TableView is asking for it, the only action is to ask for
  the item to be reloaded.
- The Switch delegate is handled at the same time as the switch cell
  configuration

Bug: None
Change-Id: I53b4ba6f95c35baa2d98980d9261045151267973
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4987550
Commit-Queue: Gauthier Ambard <gambard@chromium.org>
Reviewed-by: Louis Romero <lpromero@google.com>
Code-Coverage: findit-for-me@appspot.gserviceaccount.com <findit-for-me@appspot.gserviceaccount.com>
Cr-Commit-Position: refs/heads/main@{#1240580}