Move testable functions out of anon-namespace in form_autofill_util.h
Those functions were defined in an anonymous namespace but then a
FooForTesting version was globally defined and tested.

Since this still exposes the logic, the pattern is not particularly
useful. This CL moves the function out of the namespace and deletes
the FooForTesting functions.

This CL is a no-op and just code moving.

Bug: 1007974
Change-Id: I86fe9f357b7784dea756223ba4bf41d4ab3c8a86
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5145011
Reviewed-by: Florian Leimgruber <fleimgruber@google.com>
Auto-Submit: Jihad Hanna <jihadghanna@google.com>
Commit-Queue: Jihad Hanna <jihadghanna@google.com>
Code-Coverage: findit-for-me@appspot.gserviceaccount.com <findit-for-me@appspot.gserviceaccount.com>
Cr-Commit-Position: refs/heads/main@{#1240591}