Add feature to cancel the effect of MainThreadCompositingPriority when WebRTC is used.
There is strong evidence that MainThreadCompositingPriority causes
audio glitches and video cpu adaptation while in a WebRTC call.
Inspection of traces on Windows and MacOS suggest that the main thread
of the renderer hosting the call competes with other threads that are
critical to the quality of the video call.

Disabling MainThreadCompositingPriority entirely is not an ideal
solution, because it would cause serious regressions to Web Vitals.
This CL creates a feature to disable it only for renderers hosting a
WebRTC call. We'll look into more refined solutions over the next
quarter, but in the meantime, having a simple / low-risk solution that
can be merged to higher channels is important.

Bug: 1513904
Change-Id: I5c26fd5a0a3ca9d49344cb276e801291817b8873
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5148969
Auto-Submit: Francois Pierre Doray <fdoray@chromium.org>
Reviewed-by: Etienne Pierre-Doray <etiennep@chromium.org>
Commit-Queue: Etienne Pierre-Doray <etiennep@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1240746}