CDP: savely reset locale override
Clearing locale override unconditionally in
InspectorEmulationAgent::disable sometimes leads to a DCHECK failure as
it's not safe. What we actually want to do here is to reset the locale
override if it was previously set. To do that, resetting the unique_ptr
suffices.

R=bmeurer@chromium.org

Fixed: 1513899
Change-Id: I212465579e96b5803bc176796658219f61847587
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5147311
Auto-Submit: Yang Guo <yangguo@chromium.org>
Reviewed-by: Benedikt Meurer <bmeurer@chromium.org>
Commit-Queue: Benedikt Meurer <bmeurer@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1241109}