[Sync] Prefer server key pair in case of conflict
Prior to this CL, only the public key was updated if there is a new key
pair on the server. This CL fixes this and now both public and private
keys are updated if there is a new key on the server with the same
version. This scenario could happen in case if an old Chrome client is
upgraded to a newer version and re-generates a new key pair (since it's
not aware of the key pair on the server).

Bug: 1511180
Change-Id: Ifbbf3827fd34fa51db4b8727be162ea94119b983
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5144947
Reviewed-by: Maksim Moskvitin <mmoskvitin@google.com>
Commit-Queue: Rushan Suleymanov <rushans@google.com>
Cr-Commit-Position: refs/heads/main@{#1241139}