[iOS] Deprecate remaining UIMenuController code in OmniboxVC
In a previous CL, the OmniboxViewController was reworked to use the new
UIEditMenuInteraction object. However, when the minimum iOS deployment
target is bumped to iOS 16, a build error is generated because the old
UIMenuController object is still accessible in the code even though it
is no longer used.

In this CL, the remaining uses of UIMenuController were gated behind a
compile-time flag.

Bug: 1489281
Change-Id: I0e4adbaee05c009d799faa322a61ce81dd3bef4d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5117144
Reviewed-by: Robbie Gibson <rkgibson@google.com>
Commit-Queue: Daniel White <danieltwhite@google.com>
Code-Coverage: findit-for-me@appspot.gserviceaccount.com <findit-for-me@appspot.gserviceaccount.com>
Cr-Commit-Position: refs/heads/main@{#1241165}