[WebAuthn] Attach OAuth token to cloud enclave WebSocket connection
This retrieves an OAuth token that will allow connecting to a production
cloud enclave authenticator and puts it in the Authorization header.

This uses an EventStream to pass the token to the authenticator. It will
only be called once but since fetching the token is asynchronous this
lets us incorporate the wait into the authenticator's state machine.

Low-Coverage-Reason=Test support for enclave connections TBD

Bug: 1459620
Change-Id: Iffa842552470cbd73d7dc835eb6fed88806b2267
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5151931
Reviewed-by: Adam Langley <agl@chromium.org>
Commit-Queue: Ken Buchanan <kenrb@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1241184}