WebUI: Fix cr-slider ripples going away on every keydown
cr-slider uses a paper-ripple to indicate focus has moved onto it,
but the paper-ripple goes away as soon as you interact with the slider
using the keyboard even if focus remains on the slider. This is
caused by paper-ripple removing any existing ripples whenever its
`showAndHoldDown` method is called [1].

This CL adds a check to see if it the ripple is already shown before
calling `showAndHoldDown`.

[1] https://source.chromium.org/chromium/chromium/src/+/main:third_party/polymer/v3_0/components-chromium/paper-ripple/paper-ripple.js;drc=28d0efb352e9aed1897f88571b261696a3254306;l=119

Bug: b:315017183
Change-Id: I133504b82750b99943c32c172c166c52e81c983b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5152473
Commit-Queue: John Lee <johntlee@chromium.org>
Reviewed-by: Rebekah Potter <rbpotter@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1241215}