Improve LoginHandler (Part 7 / N)
This CL is mostly a refactor. It has a slight behavior change that
should not be noticeable to users.

(1) This CL moves logic for checking for prerender from LoginHandler to
HttpAuthCoordinator. This is a refactor and has no behavior change.

(2) This CL deletes logic that cancelled the http auth dialog if the
password manager was not available. This was added 9 years ago at
https://codereview.chromium.org/8741019/ due to a NOTREACHED seen in the
wild that could not be reproduced or otherwise diagnosed. In the current
code, this check does not make sense, as it's already possible to show
the http auth dialog without a password manager. Instead, this CL adds a
new check which is both necessary and sufficient for being able to show
an http auth dialog.

Bug: 1371177
Change-Id: I13fcd7260c59fc0d6a9c3c3add14aad66db97565
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5147846
Code-Coverage: findit-for-me@appspot.gserviceaccount.com <findit-for-me@appspot.gserviceaccount.com>
Reviewed-by: Rachel Blum <groby@chromium.org>
Commit-Queue: Erik Chen <erikchen@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1241279}