[Privacy Sandbox] Record the attestation list version that creates the sentinel file.
This is done by writing the attestation list version to the sentinel
file content when creating the sentinel file. Next time when parsing
is aborted due to the sentinel file, the version is read from its
content and reported to a histogram. The version string is converted
to an integer which is recorded to a sparse histogram.

This histogram will help us investigate the root cause for the high
percentage of parsings being aborted due to sentinel file.

Bug: 1512626, 1448875
Change-Id: I4dc8054c0dcbeaf02070f190ad2f9e41c6ddf8ef
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5132235
Reviewed-by: Theodore Olsauskas-Warren <sauski@google.com>
Commit-Queue: Xiaochen Zhou <xiaochenzh@chromium.org>
Reviewed-by: Weilun Shi <sweilun@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1241350}