Add command result payload to LogUploadEvent proto
The command result will only be uploaded to DMServer or commands that
have finished executing. For FETCH_SUPPORT_PACKET command, the command
is marked as acked when the command job has finished executing. That's
why the command result payload is not uploaded for ack'ed result. The
command will be marked as executed in the server when the LogUploadEvent
is uploaded to Reporting server. That's why a new proto field
`command_result_payload` is added to LogUploadEvent proto for the server
to attach it to command result when upload is completed.

Bug: b:317955924
Change-Id: I0653039ec78b58fb8b88a4898cc901e2b2e0a1f4
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5150735
Commit-Queue: Irem Uguz <iremuguz@google.com>
Reviewed-by: Pavol Marko <pmarko@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1241365}