[omnibox][ml] Keep the default suggestion fixed to search or non-search
go/autocomplete-controller-refactor has some details on the refactor
plan.

This is part 6 of the refactor.

ML tries to keep the ranking of searches and URLs fixed. It's allowed to
rerank URLs among themselves, but not move them relative to searches.

However, due to not considering `allowed_to_be_default`, it can cause
the default suggestion to swap from a search to a URL or vice versa when
it picks a top URL suggestion that is|isn't allowed to be default and
the traditional scoring picked a top URL suggestion that wasn't|was
allowed to be default).

This CL fixes this. If the default suggestion was a search|URL with
traditional ranking, then it will remain a search|URL with ML ranking.

The fix is in `AutocompleteController::RunBatchUrlScoringModel()`.
Before, we collected the URL scores, and redistributed them according to
ML ranking; the highest ML ranked URL would get the highest traditional
URL score. Likewise, we redistributed shortcut boosting, the N highest
ML ranked suggestions would get boosted.

With this CL, the new approach is

1. Sort the suggestions according to traditional ranking.
2. Check whether the default suggestion is a URL or search.
3. Check the # of URLs above searches (i.e. shortcut boosting) other
   than the default suggestion. Not counting the default match ensures
   the # of URLs-boosted-above-searches remains the same even if the
   traditional default match was a boosted shortcut that is no longer
   default with ML ranking.
4. If the traditional default was a URL, find the new top-ML-ranked and
   `allowed_to_be_default` match and assign it the highest traditional
   score. Picking an `allowed_to_be_default` match ensures ensures the
   default suggestion won't switch from a URL to a search.
5. Otherwise, if the traditional default was a search, cap the pool of
   scores below the default search's score; this ensures the default
   suggestion won't switch from a search to a URL.
6. Reassign shortcut boosting and relevance scores.

This change is param guarded: `stable_search_ranking`. The original
`RunBatchUrlScoringModel()` is left in the code in case we decide to
launch before we've experimented with this fix.

This CL also extracts `AutocompleteResult::Sort()` from
`AutocompleteResult::SortAndCull()` so that the former can be called
independently. This is needed to determine the default match without
wasting effort on the cull and other post processing steps. This is only
needed for the short term until we have ML search ranking, after which,
we won't need to preserve the default match type.

This CL also removes the `ml_url_scoring_rerank_final_matches_only` code
paths. They're not launch-likely and I didn't want to worry updating
that code path to fit with this CL.

Bug: 1405555
Change-Id: Iada36016ad737f9f4bd43fa9a521a833b292ff4b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5153223
Code-Coverage: findit-for-me@appspot.gserviceaccount.com <findit-for-me@appspot.gserviceaccount.com>
Reviewed-by: Angela Yoeurng <yoangela@chromium.org>
Commit-Queue: manuk hovanesian <manukh@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1241431}