[iOS] Use a non-empty string for PrimaryButton creation
This CL changes the creation of the PrimaryButton to use a non-empty
string. It allows the font to be kept into consideration when the
string is updated later.

If the string is empty, then SetConfigurationTitle is not able to get
the font, so the default font is used.

Fixed: 1502867
Change-Id: Ibf647bd0ccfe057fb144af1c68183988c3b2d248
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5147313
Reviewed-by: Joemer Ramos <joemerramos@chromium.org>
Commit-Queue: Gauthier Ambard <gambard@chromium.org>
Code-Coverage: findit-for-me@appspot.gserviceaccount.com <findit-for-me@appspot.gserviceaccount.com>
Auto-Submit: Gauthier Ambard <gambard@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1241475}