PEPC BugFix: Remove vertical white space from prompt
Secondary UI in PEPC prompts that do not have both a title and a
subtitle have more whitespace than necessary where the title should be.

To circumvent this problem, if the prompt view doesn't have a title
the subtitle will play the role of the title.

Free space is fixed in the following prompts:
*System settings view
*Show System Settings Prompt
*Previously Denied
*Previously Granted

Follow up: *Admin prompt (because it has an icon)

Bug: 146293
Change-Id: Iaa388cbbafd575aebb89fa752773d69ee2600ad7
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5148812
Reviewed-by: Ravjit Uppal <ravjit@chromium.org>
Commit-Queue: Kamila Hasanbega <hkamila@google.com>
Cr-Commit-Position: refs/heads/main@{#1241506}