[Passwords][Fixit]Display a custom string when pwd saving is managed
Android-only.

With UPM rolled out, the save password setting state is displayed
outside of Chrome, so there isn't a chance to display a "Managed by
administrator" message in the context of the toggle itself. Instead,
the message is displayed at the password manager entry point in
settings. Currently the string doesn't specify what is managed,
so this CL introduces 2 new strings that contain info about both
the setting and the state.

Bug: 1457393
Change-Id: Iff9d060a5f06dd049ce7fdd128fdad75f7262d08
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4708150
Reviewed-by: Friedrich Horschig <fhorschig@chromium.org>
Code-Coverage: findit-for-me@appspot.gserviceaccount.com <findit-for-me@appspot.gserviceaccount.com>
Commit-Queue: Ioana Pandele <ioanap@chromium.org>
Reviewed-by: Wenyu Fu <wenyufu@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1241532}