Store emoji picker global variant preferences
This CL stores the user's skin tone and gender preferences in local
storage when the flag is enabled. Individual emoji preferences are no
longer stored for emojis where the new grouped preferences apply.

New fields are also added to the existing recently-used store, to enable
users setting their global preference when selecting a variant from one
the recently used emoji. If the user has existing recently-used data,
the missing fields are filled in as emoji ordering data is loaded. There
is a very brief window after loading the emoji picker where selecting a
variant of a recently-used emoji will use the old behaviour. For future
loads this won't be an issue, because the tone and gender information is
cached in storage.

This process of filling in missing fields is potentially expensive and
happens on every load. A future CL may improve this issue by ensuring
the process only occurs once to migrate the old data. This CL would also
clean up any old individual preferences which are no longer needed.

Bug: b:317935340
Change-Id: I4233976618b50726f1ea86ff2fdf9bd55afe944f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5149252
Commit-Queue: Luke Fisk-Lennon <lukefl@google.com>
Reviewed-by: Chuong Ho <hdchuong@chromium.org>
Reviewed-by: Mehrab N <mehrab@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1241809}