[//components/content_settings] Remove Blink deps from core code
Code shared with iOS should not reference //third_party/blink (even
conditionally). Fortunately, the dependencies here are now stale. This
CL removes them.

Removing the usage of USE_BLINK from the core code of this component
would be ideal but is out of scope of this CL.

Bug: 1411765
Change-Id: I573a4e93cdbcfd34cd158ae64f496d8ed0394ea8
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5148826
Reviewed-by: Balazs Engedy <engedy@chromium.org>
Commit-Queue: Colin Blundell <blundell@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1241878}