🧇 Ensure that GetPrepopulatedTemplateURLData never gets null prefs
A previous change started making it required to pass prefs when calling
this method, but Android is the last platform where the
SearchEngineChoice feature that requires it is not enabled. In
preparation of rolling out changes related to this feature on Android
too, we close this outstanding issue.

Noting that:
1. We still have many tests passing null prefs, so this
exception is allowed for now, until we fix this in follow-ups, as
part of https://crbug.com/1499181
2. We are assuming that the service should be available when methods
are called from Java, and currently return fallback values if it's
not the case. We will move the service availability check to Java in
another follow-up (see b/318339172).

Bug: 1499181
Fixed: b:303632061
Change-Id: If1907b10df05d3f902016d911d5300f259e6865f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5001427
Reviewed-by: Boris Sazonov <bsazonov@chromium.org>
Commit-Queue: Boris Sazonov <bsazonov@chromium.org>
Reviewed-by: Orin Jaworski <orinj@chromium.org>
Auto-Submit: Nicolas Dossou-Gbété <dgn@chromium.org>
Reviewed-by: Tomasz Wiszkowski <ender@google.com>
Cr-Commit-Position: refs/heads/main@{#1241895}