[Signin] Fix race conditions in FakeAccountManagerFacade
FakeAccountManagerFacade allows reading/changing the list of accounts
from different threads. Prior to this CL, it was avoiding data races
by using a lock object. However, this was still incorrect from the
behavioral point of view - it meant that calling getCoreAccountInfos
multiple times from the same method on the main thread could produce
different results, which is not how it works in production.

This CL fixes the issue by replacing the lock object with checkUiThread
and runOnUiThreadBlocking - now mAccountHolders is only accessed on
the UI thread.

Bug: 1513740, 1493636
Change-Id: I6877cd8809f8221294c011f7cdb1cdf82c3db1dc
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5151322
Reviewed-by: Tanmoy Mollik <triploblastic@google.com>
Commit-Queue: Boris Sazonov <bsazonov@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1241903}