[Code Health][iOS] Move ios/c/b/follow to model subfolder
To simplify the folder hierarchy on iOS and have a clearer layer
separation, all the files in ios/c/b/<foo>/ will be moved to
ios/c/b/<foo>/model/ to showcase that they are model files.

This is the last part of the chain of CLs for the model/ migration:
  (DONE) #1 Migrate downstream and use forwarder where needed
         (http://crrev.com/c/5059496)
  (DONE) #2 Use model/ in internal: crrev.com/i/6706632
         (https://crrev.com/i/6082837)
THIS >>#3 Finish the migration of the not yet migrated dependencies. <<



Bug: 1319852
Change-Id: Ic2b78fdb98369d9f3269b3c7626dba9375b95adf
Low-Coverage-Reason: Code migration
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5120315
Reviewed-by: Sylvain Defresne <sdefresne@chromium.org>
Commit-Queue: Vincent Boisselle <vincb@google.com>
Cr-Commit-Position: refs/heads/main@{#1241986}