[Downloads] Consolidate and fix Download.ShowedDownloadWarning logging
Previously we only logged to this histogram when a download warning was
shown in the download bubble, a notification, or the obsolete download
shelf. Given that the download bubble is suppressed for 15s after it is
shown, it is possible for a dangerous download warning to occur but not
be shown in the download bubble. Later on, the user could go to
chrome://downloads and be shown the warning, but this wouldn't be logged
to the histogram. This causes the histogram count to be lower than the
actual number of warnings shown to the user.

This CL fixes that by logging from chrome://downloads as well. We ensure
that each warning is only logged once by checking and setting the
WasUIWarningShown bit on the DownloadItemModel.

Since the missing warnings from chrome://downloads are expected to be
a low number, we do not change the name of the histogram. This histogram
is particularly costly to rename because it contains many years of
historical data and is used to populate internal dashboards.

Also some drive-by fixes.

Bug: 1513746
Change-Id: I72b794b1a72baf02d372c2dc6720a03c34b40d35
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5147288
Reviewed-by: Daniel Rubery <drubery@chromium.org>
Commit-Queue: Lily Chen <chlily@chromium.org>
Reviewed-by: Xinghui Lu <xinghuilu@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1241995}