Reland Performance Settings Picker Tests
This is a reland of https://chromium-review.googlesource.com/c/chromium/src/+/4784086

Original IgnoreIneligibleTabs test was flaky on Linux MSAN so I re-wrote
that test and is consistently passing now.

In the original CL, I moved the the performance settings tests to the
views sub-directory because one test needed to access the tab close
button to check if the tab picker updates when a tab is closed.
However, I found a different way to close the tab without needing to
click on the close button and thus do not need to move the tests
to the views sub-directory anymore.

Patchset 1 is the original CL that was flaky and reverted.
Patchset 2 moves my changes in the reverted CL to the same file as the
existing performance settings tests. This makes it easier to see the fix
to the flaky test by comparing patchset 2 with the latest uploaded
patchset.

Change-Id: Id13ae96b5fefb9c8fe44df59404bf30b088f5369
Bug: 1513377
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5154067
Code-Coverage: findit-for-me@appspot.gserviceaccount.com <findit-for-me@appspot.gserviceaccount.com>
Reviewed-by: Eshwar Stalin <estalin@chromium.org>
Commit-Queue: Steven Luong <stluong@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1242001}